hacked together a fix for offsets caused by averaging over transition times
set the average current level caluclation to ignore the first and last
20% of each section when calculating averages for sublevels. This stops
points in the transition region from being included in the average and
improves sublevel identification for short events. It's also an ugly
hack, so make it prettier later.

+++ utils.h
@@ -41,7 +41,7 @@ typedef struct Event event;
 struct Edge
 {
     uint64_t location;
-    int type;
+    uint64_t type;
     struct Edge *next;
 };
 typedef struct Edge edge;
